update things to work with multiple courses
this updates the site to be able to render multiple different courses.

each course lives, in the same format as before, under the directory

site/content/courses/$COURSE_ID

this is set up so the output of the ocw-to-hugo script can basically be
directly copied in to site/content/courses. As we work more on that
script this will make it easy for us to test it on all the courses.

this commit also adds an extremely basic placeholder list of all the
courses which lives at `/courses`.

pr #64