update the desktop design
this does a few things, related to getting the 'course info' bar in shape.

- redo the overall layout to have even spacing around the left-side
  nav, the main content (the middle), and the course info panel
- add an .active class on the link which is currently active in the
  desktop nav and styled that so you can tell what page you're on
- re-style the course info panel in line with the designs

I also did some refactoring to reduce duplication between the
site/layout/courses/section.html and site/layout/courses/single.html
files. before they were both defining the columns and so on, which meant
that changing the design required editing both files in the same way.

what I did to fix that was to move the shared code out into
site/layout/courses/_default.html so that the other two files only have
the markup in them which is specific to them.

pr #68