add course info mobile drawer, refactor drawers in general
this adds a second drawer, on the right-hand side, to hold the course
info panel on mobile. It also adds a button for opening that drawer.

In order to get this to work I had to swap out our existing drawer
component (which was from the bootstrap material design package). I
tried out a few options before settling on this one:
https://github.com/iamphill/Bootstrap-Offcanvas

we're now only using the BMD package for styling the search bar up at
the top. when we re-do that we'll remove it.

pr #74
https://trello.com/c/vjF0VNOk/26-mobile-course-navigation-drawer
https://trello.com/c/3ajMJlbT/25-mobile-course-info-drawer