webpack config tweaks
two tweaks to improve webpack user experience in development:

first, I turned off webpack watching the `/dist` folder, since it gets
too heavy and causes webpack to sort of blow up. this is set via the
`watchContentBase` option in the webpack config. We're basically setting
hugo to a watch mode, where it doesn't serve anything but simply writes
updated files to the `dist` directory. Then we use webpack's dev server
to serve the HTML for development.

this gets problematic because if you update a template that is used on a
lot of pages, like layouts/courses/_baseof.html, for instance, hugo will
write something like 2000 html files into `/dist`. Webpack will see all
of these file changes and basically blow up with the `watchContentBase`
option.

I also got our webpack HMR setup in shape. We were missing two options,
the first being setting a `publicPath` in `output` in the dev config,
and the other passing `hmr: true` to the MiniCSSExtractPlugin config in
development. The combination of these allows hot-reloading of CSS, which
is very nice!

pr #78