switch from using webpack as main server to using hugo
tooling change, this sets us up to use the hugo server instead of using
webpack to serve html.

this is nice, because hugo server has a built-in live-reloading thing
which is very fast. it seems to work just fine alongside the webpack
hot-reloading as well.

you can use it either in the docker container or just run it directly on
your laptop.

pr #82