add stylesheet switching system
this adds a little system for picking between a few different
stylesheets to apply in the app.

I added a little drop-down in the upper right hand corner of the page.
selecting an option in the dropdown sets a class on the html element,
which can then by used to style the page differently. I included one
very ugly test theme just to demonstrate it's working.

pr #93