add search functionality to ocw-next
this adds a search feature to ocw-next. the search is provided by an
external service, namely, the existing course search in MIT open.  In
order to use the feature an environment variable needs to be set, see
the README for details.

the feature is built with React, our first such feature in this app, so
naturally implementing it required some tooling and setup changes. I
revamped slightly the webpack config, and also added something to help
us keep track of our bundle size as we're working. check the README for
details of that too :)

I also changed our tests in this repo over to use Jest instead of mocha.
I'm finding it pretty nice so far, but if we decide we don't like it the
syntax is largely compatible and I think it would not be much work to
switch back.

closes #158
pr #172