fix search environment variable access
this changes the config for the dotenv-webpack plugin which we're using
to pull environment variables into webpack. by default it only pulls
variables from the `.env` file and not from the environment. for local
development we want to use a `.env` file, but for the CI netlify builds
we can only use the environment to configure things.

so, to fix the situation, this PR sets the `systemvars` option in the
dotenv-webpack plugin config to true, which will mean that the bundle
will pick up and include referenced variables from the environment and
not just from `.env`.

pr #179