remove automatic search behavior
this removes the behavior where as the user was typing we would fire off
a debounced (300ms) callback to automatically run the search.

Now, instead, the user must manually trigger the search by pressing
'enter' or clicking on the little spyglass.

I also added the latest version of our eslint rules, and fixed issues
found after the upgrade.

pr #190