remove moment.js from dependencies for this project
this removes moment, which we weren't actually using. I pulled it in as
part of some code that was pulled over from Open, but the particular
code that depended on moment we haven't needed yet. So I decided to just
delete it, and if / when we need equivalent functionality over here we
can implement it without moment.

pr #206