set partial collapse list to reveal 4 items
this sets the collapsible list to reveal 4 items by default instead of three items as it did
previously. it reveals three 'full' items, and partially reveals the fourth, giving the user
a visual hint that they can expand the list to see more. if the list has four items or
fewer we just show the whole list.

closes #220 
pr #229