downgrade react-dotdotdot
this downgrades the react-dotdotdot package to the same version we're
using in Open. The newer version doesn't seem to recalculate whether it
should truncate or not correctly (only on re-render, not if a hover
state changes, for instance).

The older version does better, and solves
one of the issues listed in #226, namely that our facet labels were
getting pushed onto two lines when the user hovered over them, because
we increase the font size.

part of #226
pr #234