fix issue with loading placeholder
fix an issue where the loading placeholder was always being shown below
the search results. This was happening because I was inadvertantly
passing the same <Loader /> component in to the <InfiniteScroll>
component and using it as a loading placeholder.

To fix it, I brought the spinner we had back to use for showing at the
end of the search results and keep the loading placeholder just for the
initial load.

pr #236