add level facet to search
This adds support for the 'level' facet, by upgrading
course-search-utils and making some changes to the code used to generate
the ES query. It works basically like our other facets, with the caveat
that the data is stored on course runs, not on the courses themselves,
so we have to do a nested query for it.

closes #216
pr #267