fix issue with title width
the whole title area had `cursor: pointer` but the anchor tag didn't
have `width: 100%` on it so there was a confusing area on the search
result title where it looked, to the user, like it should be clickable
but it in fact was not.

this fixes that up so that the anchor tag fills the title div.

closes #270
pr #272