link from course info to search
this re-writes the href of the links in the course info side panel to
link to the search with the relevant parameter and value pre-selected.
this is so you can, for instance, click 'Undergraduate' in the panel
to go to "/search/?l=Undergraduate".

this adds such links for the course department and level. in order to
add instructor we need to make some larger changes to the search code,
and for course topics we need to make a decision about how to handle the
'third level' topics on OCW courses. for more on that see this comment:
https://github.com/mitodl/hugo-course-publisher/pull/281#issuecomment-719547924

part of #273
pr #281