slightly reduce bundle size, style tables correctly
I wanted to remove the bootstrap-material-design package since it was
around 20kb of JS that we're not using. After removing it, Matt pointed
out that tables looked slightly different, so I went ahead and restyled
tables to follow the current designs on invision (which current master
wasn't implementing either).

Anyhow, this reduced the size of the gzipped JS bundle by around 20kb,
which isn't a _huge_ amount but is something.

pr #291