update styling of the home page to match designs
this updates the styling of the home page to match our latest mocks,
integrating a big banner image, some new copy, and a new top nav /
header component.

I also figured out a way to remove the react component being used for
the homepage search box, replacing it with a <form> and two <input>s.

part of #314
pr #343