expand example_courses.json
thought we should add a series of courses which have had some issue at
some point, that way when we're developing the site we're exercising
more of the things we've implemented.

As we find issues with courses we should make a practice of adding a
representative course to `example_courses.json`.

pr #390