add partialCached in a few places
this caches the rendered output of partials between .md files, allowing
us to save some resources. This should make the build a bit faster.

we get a small boost from caching the course info partials (I added one
for the desktop course into to make it easier to cache). Both are cached
under the course ID as a cache key, because the rendered templates
should be the same for every page in the course.

also added caching to a few partials which don't change no matter the
course they're rendered on, like the header partial and so on.

these changes give a roughly ~15% speedup in hugo build time - times
varied a little bit, but it took the build from ~12-13s to ~10 (this is
running off of the courses in the example_courses.json file). We should
see an equivalent drop in build times for the whole course set, I should
think. for details see some debug output on the PR.

This still doesn't touch the real hugo build performance issue, which is
the rendering of the nav. We're doing a complicated recursive thing, and
rendering those course navs occupies the vast majority of template
processing time.

pr #393