Deploy to netlify with github actions, add google lighthouse
This adds two github actions, one to deploy PR and CI builds to Netlify
and another to then run google lighthouse on the deployed site. We have
to do both of these at the same time because we need to guarantee that
the netlify build is done before we run the lighthouse check.

For now I set up lighthouse to just write a comment on the PR with the
scores. We can circle back later to add something more complicated if we
want to.

closes #367
pr #430